Refactor and separate adminBro files to multiple file
