+++ src/server/adminBro/tabs/mainCategoryTab.ts
@@ -1,0 +1,59 @@
+import { MainCategory } from '../../../entities/mainCategory';
+import {
+  canAccessMainCategoryAction,
+  ResourceActions,
+} from '../adminBroPermissions';
+
+export const mainCategoryTab = {
+  resource: MainCategory,
+  options: {
+    actions: {
+      delete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessMainCategoryAction({ currentAdmin }, ResourceActions.DELETE),
+      },
+      new: {
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessMainCategoryAction({ currentAdmin }, ResourceActions.NEW),
+      },
+      edit: {
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessMainCategoryAction({ currentAdmin }, ResourceActions.EDIT),
+      },
+      bulkDelete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessMainCategoryAction(
+            { currentAdmin },
+            ResourceActions.BULK_DELETE,
+          ),
+      },
+    },
+    properties: {
+      id: {
+        isVisible: {
+          list: true,
+          filter: true,
+          show: true,
+          edit: false,
+          new: false,
+        },
+      },
+      banner: {
+        isVisible: true,
+      },
+      slug: {
+        isVisible: true,
+      },
+      title: {
+        isVisible: true,
+      },
+      description: {
+        isVisible: true,
+      },
+    },
+  },
+};
