+++ src/server/adminBro/tabs/projectStatusReasonTab.ts
@@ -1,0 +1,43 @@
+import { ProjectStatusReason } from '../../../entities/projectStatusReason';
+import {
+  canAccessProjectStatusReasonAction,
+  ResourceActions,
+} from '../adminBroPermissions';
+
+export const projectStatusReasonTab = {
+  resource: ProjectStatusReason,
+  options: {
+    actions: {
+      new: {
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusReasonAction(
+            { currentAdmin },
+            ResourceActions.NEW,
+          ),
+      },
+      edit: {
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusReasonAction(
+            { currentAdmin },
+            ResourceActions.EDIT,
+          ),
+      },
+      delete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusReasonAction(
+            { currentAdmin },
+            ResourceActions.DELETE,
+          ),
+      },
+      bulkDelete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusReasonAction(
+            { currentAdmin },
+            ResourceActions.BULK_DELETE,
+          ),
+      },
+    },
+  },
+};
