+++ src/server/adminBro/tabs/projectStatusTab.ts
@@ -1,0 +1,38 @@
+import { ProjectStatus } from '../../../entities/projectStatus';
+import {
+  canAccessProjectStatusAction,
+  ResourceActions,
+} from '../adminBroPermissions';
+
+export const projectStatusTab = {
+  resource: ProjectStatus,
+  options: {
+    actions: {
+      delete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusAction(
+            { currentAdmin },
+            ResourceActions.DELETE,
+          ),
+      },
+      new: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusAction({ currentAdmin }, ResourceActions.NEW),
+      },
+      edit: {
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusAction({ currentAdmin }, ResourceActions.EDIT),
+      },
+      bulkDelete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectStatusAction(
+            { currentAdmin },
+            ResourceActions.BULK_DELETE,
+          ),
+      },
+    },
+  },
+};
