+++ src/server/adminBro/tabs/projectsTab.ts
@@ -1,0 +1,1106 @@
+import {
+  Project,
+  ProjectUpdate,
+  ProjStatus,
+  ReviewStatus,
+  RevokeSteps,
+} from '../../../entities/project';
+import AdminBro from 'adminjs';
+import {
+  canAccessProjectAction,
+  ResourceActions,
+} from '../adminBroPermissions';
+import {
+  findProjectById,
+  findProjectsByIdArray,
+} from '../../../repositories/projectRepository';
+import { NOTIFICATIONS_EVENT_NAMES } from '../../../analytics/analytics';
+import { HISTORY_DESCRIPTIONS } from '../../../entities/projectStatusHistory';
+import { getNotificationAdapter } from '../../../adapters/adaptersFactory';
+import { changeUserBoostingsAfterProjectCancelled } from '../../../services/powerBoostingService';
+import { refreshUserProjectPowerView } from '../../../repositories/userProjectPowerViewRepository';
+import {
+  refreshProjectFuturePowerView,
+  refreshProjectPowerView,
+} from '../../../repositories/projectPowerViewRepository';
+import { redis } from '../../../redis';
+import { SelectQueryBuilder } from 'typeorm';
+import { logger } from '../../../utils/logger';
+import {
+  ActionResponse,
+  After,
+} from 'adminjs/src/backend/actions/action.interface';
+import { RecordJSON } from 'adminjs/src/frontend/interfaces/record-json.interface';
+import { findSocialProfilesByProjectId } from '../../../repositories/socialProfileRepository';
+import { findProjectUpdatesByProjectId } from '../../../repositories/projectUpdateRepository';
+import {
+  AdminBroContextInterface,
+  AdminBroProjectsQuery,
+  AdminBroRequestInterface,
+  headers,
+} from '../adminBro-types';
+import { ProjectStatus } from '../../../entities/projectStatus';
+import { messages } from '../../../utils/messages';
+import {
+  addSheetWithRows,
+  projectExportSpreadsheet,
+} from '../../../services/googleSheets';
+import { NETWORKS_IDS_TO_NAME } from '../../../provider';
+import {
+  getVerificationFormByProjectId,
+  makeFormDraft,
+  makeFormVerified,
+} from '../../../repositories/projectVerificationRepository';
+import { FeaturedUpdate } from '../../../entities/featuredUpdate';
+
+// add queries depending on which filters were selected
+export const buildProjectsQuery = (
+  queryStrings: AdminBroProjectsQuery,
+): SelectQueryBuilder<Project> => {
+  const query = Project.createQueryBuilder('project')
+    .leftJoinAndSelect('project.addresses', 'addresses')
+    .leftJoinAndSelect('project.adminUser', 'adminUser')
+    .where('addresses.isRecipient = true')
+    .addOrderBy('addresses.networkId', 'ASC');
+
+  if (queryStrings.title)
+    query.andWhere('project.title ILIKE :title', {
+      title: `%${queryStrings.title}%`,
+    });
+
+  if (queryStrings.slug)
+    query.andWhere('project.slug ILIKE :slug', { title: queryStrings.slug });
+
+  if (queryStrings.verified)
+    query.andWhere('project.verified = :verified', {
+      verified: queryStrings.verified === 'true',
+    });
+
+  if (queryStrings.isImported)
+    query.andWhere('project.isImported = :isImported', {
+      isImported: queryStrings.isImported === 'true',
+    });
+
+  // if (queryStrings.listed)
+  //   query.andWhere('project.listed = :listed', {
+  //     listed: queryStrings.listed === 'true',
+  //   });
+  //
+  if (queryStrings.reviewStatus)
+    query.andWhere('project.reviewStatus = :reviewStatus', {
+      reviewStatus: queryStrings.reviewStatus,
+    });
+
+  if (queryStrings.statusId)
+    query.andWhere('project."statusId" = :statusId', {
+      statusId: queryStrings.statusId,
+    });
+
+  if (queryStrings['creationDate~~from'])
+    query.andWhere('project."creationDate" >= :createdFrom', {
+      createdFrom: queryStrings['creationDate~~from'],
+    });
+
+  if (queryStrings['creationDate~~to'])
+    query.andWhere('project."creationDate" <= :createdTo', {
+      createdTo: queryStrings['creationDate~~to'],
+    });
+
+  if (queryStrings['updatedAt~~from'])
+    query.andWhere('project."updatedAt" >= :updatedFrom', {
+      updatedFrom: queryStrings['updatedAt~~from'],
+    });
+
+  if (queryStrings['updatedAt~~to'])
+    query.andWhere('project."updatedAt" <= :updatedTo', {
+      updatedTo: queryStrings['updatedAt~~to'],
+    });
+
+  return query;
+};
+
+export const addFeaturedProjectUpdate = async (
+  context: AdminBroContextInterface,
+  request,
+) => {
+  const { records } = context;
+  try {
+    const updateIds = request?.query?.recordIds
+      ?.split(',')
+      ?.map(id => Number(id));
+    const projectUpdates = await ProjectUpdate.createQueryBuilder(
+      'projectUpdate',
+    )
+      .innerJoinAndMapOne(
+        'projectUpdate.project',
+        Project,
+        'project',
+        'project.id = projectUpdate.projectId AND projectUpdate.isMain = false',
+      )
+      .where('projectUpdate.id IN (:...updateIds)', { updateIds })
+      .andWhere('projectUpdate.isMain = false')
+      .andWhere(
+        `project.statusId = :projectStatus AND project.reviewStatus = :reviewStatus`,
+        {
+          projectStatus: ProjStatus.active,
+          reviewStatus: ReviewStatus.Listed,
+        },
+      )
+      .getMany();
+
+    for (const update of projectUpdates) {
+      const featured = await FeaturedUpdate.createQueryBuilder('featuredUpdate')
+        .where('featuredUpdate.projectId = :projectId', {
+          projectId: update!.project!.id,
+        })
+        .getOne();
+
+      if (featured) continue; // ignore if already project featured
+
+      const featuredProject = FeaturedUpdate.create({
+        projectUpdateId: update.id,
+        projectId: update.project!.id,
+      });
+
+      await featuredProject.save();
+    }
+  } catch (error) {
+    logger.error('addFeaturedProjectUpdate error', error);
+    throw error;
+  }
+  return {
+    redirectUrl: '/admin/resources/ProjectUpdate',
+    records: records.map(record => {
+      record.toJSON(context.currentAdmin);
+    }),
+    notice: {
+      message: `ProjectUpdates(s) successfully added to the featured Project list`,
+      type: 'success',
+    },
+  };
+};
+
+export const verifyProjects = async (
+  context: AdminBroContextInterface,
+  request: AdminBroRequestInterface,
+  verified: boolean = true,
+  revokeBadge: boolean = false,
+) => {
+  const { records, currentAdmin } = context;
+  // prioritize revokeBadge
+  const verificationStatus = revokeBadge ? false : verified;
+  try {
+    const projectIds = request?.query?.recordIds
+      ?.split(',')
+      ?.map(strId => Number(strId)) as number[];
+    const projectsBeforeUpdating = await findProjectsByIdArray(projectIds);
+    const updateParams = { verified: verificationStatus };
+
+    if (verificationStatus) {
+      await Project.query(`
+        UPDATE project
+        SET "verificationStatus" = NULL
+        WHERE id IN (${request?.query?.recordIds})
+      `);
+    }
+
+    const projects = await Project.createQueryBuilder('project')
+      .update<Project>(Project, updateParams)
+      .where('project.id IN (:...ids)')
+      .setParameter('ids', projectIds)
+      .returning('*')
+      .updateEntity(true)
+      .execute();
+
+    for (const project of projects.raw) {
+      if (
+        projectsBeforeUpdating.find(p => p.id === project.id)?.verified ===
+        verificationStatus
+      ) {
+        logger.info('verifying/unVerifying project but no changes happened', {
+          projectId: project.id,
+          verificationStatus,
+        });
+        // if project.verified have not changed, so we should not execute rest of the codes
+        continue;
+      }
+      await Project.addProjectStatusHistoryRecord({
+        project,
+        status: project.status,
+        userId: currentAdmin.id,
+        description: verified
+          ? HISTORY_DESCRIPTIONS.CHANGED_TO_VERIFIED
+          : HISTORY_DESCRIPTIONS.CHANGED_TO_UNVERIFIED,
+      });
+      const projectWithAdmin = (await findProjectById(project.id)) as Project;
+
+      if (revokeBadge) {
+        projectWithAdmin.verificationStatus = RevokeSteps.Revoked;
+        await projectWithAdmin.save();
+        await getNotificationAdapter().projectBadgeRevoked({
+          project: projectWithAdmin,
+        });
+      } else if (verificationStatus) {
+        await getNotificationAdapter().projectVerified({
+          project: projectWithAdmin,
+        });
+      } else {
+        await getNotificationAdapter().projectUnVerified({
+          project: projectWithAdmin,
+        });
+      }
+
+      const verificationForm = await getVerificationFormByProjectId(project.id);
+      if (verificationForm) {
+        if (verificationStatus) {
+          await makeFormVerified({
+            formId: verificationForm.id,
+            adminId: currentAdmin.id,
+          });
+        } else {
+          await makeFormDraft({
+            formId: verificationForm.id,
+            adminId: currentAdmin.id,
+          });
+        }
+      }
+    }
+
+    await Promise.all([
+      refreshUserProjectPowerView(),
+      refreshProjectPowerView(),
+      refreshProjectFuturePowerView(),
+    ]);
+  } catch (error) {
+    logger.error('verifyProjects() error', error);
+    throw error;
+  }
+  return {
+    redirectUrl: '/admin/resources/Project',
+    records: records.map(record => {
+      record.toJSON(context.currentAdmin);
+    }),
+    notice: {
+      message: `Project(s) successfully ${
+        verificationStatus ? 'verified' : 'unverified'
+      }`,
+      type: 'success',
+    },
+  };
+};
+
+export const updateStatusOfProjects = async (
+  context: AdminBroContextInterface,
+  request: AdminBroRequestInterface,
+  status,
+) => {
+  const { records, currentAdmin } = context;
+  try {
+    const projectIds = request?.query?.recordIds
+      ?.split(',')
+      ?.map(strId => Number(strId)) as number[];
+    const projectsBeforeUpdating = await findProjectsByIdArray(projectIds);
+    const projectStatus = await ProjectStatus.findOne({
+      where: { id: status },
+    });
+    if (projectStatus) {
+      const updateData: any = { status: projectStatus };
+      if (status === ProjStatus.cancelled || status === ProjStatus.deactive) {
+        updateData.verified = false;
+        updateData.listed = false;
+        updateData.reviewStatus = ReviewStatus.NotListed;
+      }
+      const projects = await Project.createQueryBuilder('project')
+        .update<Project>(Project, updateData)
+        .where('project.id IN (:...ids)')
+        .setParameter('ids', projectIds)
+        .returning('*')
+        .updateEntity(true)
+        .execute();
+
+      for (const project of projects.raw) {
+        if (
+          projectsBeforeUpdating.find(p => p.id === project.id)?.statusId ===
+          projectStatus.id
+        ) {
+          logger.info('Changing project status but no changes happened', {
+            projectId: project.id,
+            projectStatus,
+          });
+          // if project.listed have not changed, so we should not execute rest of the codes
+          continue;
+        }
+        await Project.addProjectStatusHistoryRecord({
+          project,
+          status: projectStatus,
+          userId: currentAdmin.id,
+        });
+        const projectWithAdmin = (await findProjectById(project.id)) as Project;
+        if (status === ProjStatus.cancelled) {
+          await getNotificationAdapter().projectCancelled({
+            project: projectWithAdmin,
+          });
+          await changeUserBoostingsAfterProjectCancelled({
+            projectId: project.id,
+          });
+        } else if (status === ProjStatus.active) {
+          await getNotificationAdapter().projectReactivated({
+            project: projectWithAdmin,
+          });
+        } else if (status === ProjStatus.deactive) {
+          await getNotificationAdapter().projectDeactivated({
+            project: projectWithAdmin,
+          });
+        }
+      }
+      await Promise.all([
+        refreshUserProjectPowerView(),
+        refreshProjectFuturePowerView(),
+        refreshProjectPowerView(),
+      ]);
+    }
+  } catch (error) {
+    throw error;
+  }
+  return {
+    redirectUrl: '/admin/resources/Project',
+    records: records.map(record => {
+      record.toJSON(context.currentAdmin);
+    }),
+    notice: {
+      message: messages.PROJECT_STATUS_UPDATED_SUCCESSFULLY,
+      type: 'success',
+    },
+  };
+};
+
+export const setSocialProfiles: After<ActionResponse> = async (
+  response,
+  request,
+  context,
+) => {
+  const record: RecordJSON = response.record || {};
+  // both cases for projectVerificationForms and projects' ids
+  const projectId = record.params.projectId || record.params.id;
+
+  const socials = await findSocialProfilesByProjectId({ projectId });
+  const projectUpdates = await findProjectUpdatesByProjectId(projectId);
+  const project = await findProjectById(projectId);
+  const adminBroBaseUrl = process.env.SERVER_URL;
+  response.record = {
+    ...record,
+    params: {
+      ...record.params,
+      projectUrl: `${process.env.GIVETH_IO_DAPP_BASE_URL}/project/${
+        project!.slug
+      }`,
+      socials,
+      projectUpdates,
+      adminBroBaseUrl,
+    },
+  };
+  return response;
+};
+
+const sendProjectsToGoogleSheet = async (
+  projects: Project[],
+): Promise<void> => {
+  const spreadsheet = await projectExportSpreadsheet();
+
+  // parse data and set headers
+  const projectRows = projects.map((project: Project) => {
+    const projectAddresses = project.addresses || [];
+
+    return {
+      id: project.id,
+      title: project.title,
+      slug: project.slug,
+      admin: project.admin,
+      creationDate: project.creationDate,
+      updatedAt: project.updatedAt,
+      impactLocation: project.impactLocation || '',
+      walletAddress: project.walletAddress,
+      statusId: project.statusId,
+      qualityScore: project.qualityScore,
+      verified: Boolean(project.verified),
+      listed: Boolean(project.listed),
+      reviewStatus: project.reviewStatus,
+      totalDonations: project.totalDonations,
+      totalProjectUpdates: project.totalProjectUpdates,
+      website: project.website || '',
+      email: project?.adminUser?.email || '',
+      firstWalletAddress: projectAddresses![0]?.address,
+      firstWalletAddressNetwork:
+        NETWORKS_IDS_TO_NAME[projectAddresses![0]?.networkId] || '',
+      secondWalletAddress: projectAddresses![1]?.address || '',
+      secondWalletAddressNetwork:
+        NETWORKS_IDS_TO_NAME[projectAddresses![1]?.networkId] || '',
+    };
+  });
+
+  await addSheetWithRows(spreadsheet, headers, projectRows);
+};
+
+export const listDelist = async (
+  context: AdminBroContextInterface,
+  request,
+  reviewStatus: ReviewStatus = ReviewStatus.Listed,
+) => {
+  const { records, currentAdmin } = context;
+  let listed;
+  switch (reviewStatus) {
+    case ReviewStatus.Listed:
+      listed = true;
+      break;
+    case ReviewStatus.NotListed:
+      listed = false;
+      break;
+    case ReviewStatus.NotReviewed:
+      listed = null;
+      break;
+    default:
+  }
+  try {
+    const projectIds = request?.query?.recordIds
+      ?.split(',')
+      ?.map(strId => Number(strId)) as number[];
+    const projectsBeforeUpdating = await findProjectsByIdArray(projectIds);
+    const projects = await Project.createQueryBuilder('project')
+      .update<Project>(Project, { reviewStatus, listed })
+      .where('project.id IN (:...ids)')
+      .setParameter('ids', projectIds)
+      .returning('*')
+      .updateEntity(true)
+      .execute();
+    for (const project of projects.raw) {
+      if (
+        projectsBeforeUpdating.find(p => p.id === project.id)?.reviewStatus ===
+        reviewStatus
+      ) {
+        logger.info('listing/uListing project but no changes happened', {
+          projectId: project.id,
+          reviewStatus,
+        });
+        // if project.listed have not changed, so we should not execute rest of the codes
+        continue;
+      }
+      await Project.addProjectStatusHistoryRecord({
+        project,
+        status: project.status,
+        userId: currentAdmin.id,
+        description:
+          reviewStatus === ReviewStatus.Listed
+            ? HISTORY_DESCRIPTIONS.CHANGED_TO_LISTED
+            : HISTORY_DESCRIPTIONS.CHANGED_TO_UNLISTED,
+      });
+      const projectWithAdmin = (await findProjectById(project.id)) as Project;
+      if (reviewStatus === ReviewStatus.Listed) {
+        await getNotificationAdapter().projectListed({
+          project: projectWithAdmin,
+        });
+      } else {
+        await getNotificationAdapter().projectDeListed({
+          project: projectWithAdmin,
+        });
+      }
+    }
+  } catch (error) {
+    logger.error('listDelist error', error);
+    throw error;
+  }
+  return {
+    redirectUrl: '/admin/resources/Project',
+    records: records.map(record => {
+      record.toJSON(context.currentAdmin);
+    }),
+    notice: {
+      message: `Project(s) review status successfully changed to ${reviewStatus}`,
+      type: 'success',
+    },
+  };
+};
+
+export const exportProjectsWithFiltersToCsv = async (
+  _request: AdminBroRequestInterface,
+  _response,
+  context: AdminBroContextInterface,
+) => {
+  try {
+    const { records } = context;
+    const rawQueryStrings = await redis.get(
+      `adminbro_${context.currentAdmin.id}_qs`,
+    );
+    const queryStrings = rawQueryStrings ? JSON.parse(rawQueryStrings) : {};
+    const projectsQuery = buildProjectsQuery(queryStrings);
+    const projects = await projectsQuery.getMany();
+
+    await sendProjectsToGoogleSheet(projects);
+
+    return {
+      redirectUrl: '/admin/resources/Project',
+      records,
+      notice: {
+        message: `Project(s) successfully exported`,
+        type: 'success',
+      },
+    };
+  } catch (e) {
+    return {
+      redirectUrl: '/admin/resources/Project',
+      record: {},
+      notice: {
+        message: e.message,
+        type: 'danger',
+      },
+    };
+  }
+};
+export const projectsTab = {
+  resource: Project,
+  options: {
+    properties: {
+      id: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+      },
+      changeId: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+      },
+      organizationId: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      statusId: {
+        isVisible: { list: true, filter: true, show: true, edit: true },
+      },
+      socials: {
+        type: 'mixed',
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+          new: false,
+        },
+        components: {
+          show: AdminBro.bundle('./components/VerificationFormSocials'),
+        },
+      },
+      adminUserId: {
+        isVisible: {
+          list: true,
+          filter: false,
+          show: true,
+          edit: false,
+          new: false,
+        },
+      },
+      contacts: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: false,
+          edit: false,
+          new: false,
+        },
+      },
+      qualityScore: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      verified: {
+        isVisible: {
+          list: true,
+          filter: true,
+          show: true,
+          edit: true,
+          new: false,
+        },
+      },
+      verificationStatus: {
+        availableValues: [
+          { value: 'reminder', label: 'Reminder' },
+          { value: 'warning', label: 'Warning' },
+          { value: 'lastChance', label: 'Last Chance' },
+          { value: 'upForRevoking', label: 'Ready for Revoking' },
+          { value: 'revoked', label: 'Revoked' },
+        ],
+        isVisible: {
+          list: false,
+          filter: true,
+          show: true,
+          edit: true,
+          new: false,
+        },
+      },
+      totalDonations: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+      },
+      totalTraceDonations: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      admin: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      description: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: true,
+        },
+      },
+      slug: {
+        isVisible: { list: true, filter: true, show: true, edit: true },
+      },
+      projectUrl: {
+        type: 'mixed',
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+        components: {
+          show: AdminBro.bundle('./components/ClickableLink'),
+        },
+      },
+      organisationId: {
+        isVisible: false,
+      },
+      coOrdinates: {
+        isVisible: false,
+      },
+      image: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      givingBlocksId: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      traceCampaignId: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      website: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      youtube: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      balance: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+      },
+      totalProjectUpdates: {
+        isVisible: false,
+      },
+      giveBacks: {
+        isVisible: false,
+      },
+      stripeAccountId: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+      },
+      isImported: {
+        isVisible: {
+          list: false,
+          filter: true,
+          show: true,
+          edit: false,
+        },
+      },
+      totalReactions: {
+        isVisible: false,
+      },
+      walletAddress: {
+        isVisible: { list: false, filter: true, show: true, edit: true },
+      },
+      impactLocation: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      descriptionSummary: {
+        isVisible: { list: false, filter: false, show: true, edit: true },
+      },
+      slugHistory: {
+        isVisible: false,
+      },
+      addresses: {
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+          new: false,
+        },
+        components: {
+          show: AdminBro.bundle('./components/ListProjectAddresses'),
+        },
+      },
+      listed: {
+        isVisible: false,
+        // components: {
+        //   filter: AdminBro.bundle('./components/FilterListedComponent'),
+        // },
+      },
+      reviewStatus: {
+        isVisible: {
+          show: true,
+          list: true,
+          edit: true,
+          filter: true,
+        },
+      },
+      projectUpdates: {
+        type: 'mixed',
+        isVisible: {
+          list: false,
+          filter: false,
+          show: true,
+          edit: false,
+        },
+        components: {
+          show: AdminBro.bundle('./components/ProjectUpdates'),
+        },
+      },
+      adminBroBaseUrl: {
+        type: 'string',
+        isVisible: {
+          list: false,
+          filter: false,
+          show: false,
+          edit: false,
+        },
+      },
+    },
+    actions: {
+      delete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction({ currentAdmin }, ResourceActions.DELETE),
+      },
+      new: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction({ currentAdmin }, ResourceActions.NEW),
+      },
+      bulkDelete: {
+        isVisible: false,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction({ currentAdmin }, ResourceActions.BULK_DELETE),
+      },
+      show: {
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction({ currentAdmin }, ResourceActions.SHOW),
+        after: setSocialProfiles,
+      },
+      edit: {
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction({ currentAdmin }, ResourceActions.EDIT),
+        before: async (
+          request: AdminBroRequestInterface,
+          response,
+          context: AdminBroContextInterface,
+        ) => {
+          const { verified, reviewStatus } = request.payload;
+          const statusChanges: string[] = [];
+          if (request?.payload?.id) {
+            const project = await findProjectById(Number(request.payload.id));
+            if (
+              project &&
+              Number(request?.payload?.statusId) !== project?.status?.id
+            ) {
+              switch (Number(request?.payload?.statusId)) {
+                case ProjStatus.active:
+                  statusChanges.push(
+                    NOTIFICATIONS_EVENT_NAMES.PROJECT_ACTIVATED,
+                  );
+                  break;
+                case ProjStatus.deactive:
+                  statusChanges.push(
+                    NOTIFICATIONS_EVENT_NAMES.PROJECT_DEACTIVATED,
+                  );
+                  break;
+                case ProjStatus.cancelled:
+                  statusChanges.push(
+                    NOTIFICATIONS_EVENT_NAMES.PROJECT_CANCELLED,
+                  );
+                  break;
+              }
+            }
+            if (project?.verified && !verified) {
+              statusChanges.push(NOTIFICATIONS_EVENT_NAMES.PROJECT_UNVERIFIED);
+            }
+            if (!project?.verified && verified) {
+              statusChanges.push(NOTIFICATIONS_EVENT_NAMES.PROJECT_VERIFIED);
+            }
+            if (
+              project?.reviewStatus === ReviewStatus.Listed &&
+              reviewStatus === ReviewStatus.NotListed
+            ) {
+              statusChanges.push(NOTIFICATIONS_EVENT_NAMES.PROJECT_UNLISTED);
+            }
+            if (
+              project?.reviewStatus !== ReviewStatus.Listed &&
+              reviewStatus === ReviewStatus.Listed
+            ) {
+              statusChanges.push(NOTIFICATIONS_EVENT_NAMES.PROJECT_LISTED);
+            }
+            if (
+              project &&
+              (project?.reviewStatus === ReviewStatus.Listed ||
+                project?.reviewStatus === ReviewStatus.NotListed) &&
+              reviewStatus === ReviewStatus.NotReviewed
+            ) {
+              statusChanges.push(
+                NOTIFICATIONS_EVENT_NAMES.PROJECT_NOT_REVIEWED,
+              );
+            }
+
+            // We put these status changes in payload, so in after hook we would know to send notification for users
+            request.payload.statusChanges = statusChanges.join(',');
+          }
+          return request;
+        },
+        after: async (
+          request: AdminBroRequestInterface,
+          response,
+          context: AdminBroContextInterface,
+        ) => {
+          const { currentAdmin } = context;
+          const project = await Project.findOne({
+            where: { id: request?.record?.id },
+          });
+          if (project) {
+            // Not required for now
+            // Project.notifySegment(project, SegmentEvents.PROJECT_EDITED);
+
+            // As we don't want fields be changed (listed, verified, ..), I just added new status and a description that project has been edited
+            await Project.addProjectStatusHistoryRecord({
+              project,
+              status: project.status,
+              userId: currentAdmin.id,
+              description: HISTORY_DESCRIPTIONS.HAS_BEEN_EDITED,
+            });
+            const statusChanges =
+              request?.record?.params?.statusChanges?.split(',') || [];
+
+            const eventAndHandlers = [
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_VERIFIED,
+                handler: getNotificationAdapter().projectVerified,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_UNVERIFIED,
+                handler: getNotificationAdapter().projectUnVerified,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_LISTED,
+                handler: getNotificationAdapter().projectListed,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_UNLISTED,
+                handler: getNotificationAdapter().projectDeListed,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_ACTIVATED,
+                handler: getNotificationAdapter().projectReactivated,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_DEACTIVATED,
+                handler: getNotificationAdapter().projectDeactivated,
+              },
+              {
+                event: NOTIFICATIONS_EVENT_NAMES.PROJECT_CANCELLED,
+                handler: getNotificationAdapter().projectCancelled,
+              },
+            ];
+
+            eventAndHandlers.forEach(eventHandler => {
+              if (statusChanges?.includes(eventHandler.event)) {
+                // Don't put await before that intentionally to not block admin panel response with that
+                eventHandler.handler({ project });
+              }
+            });
+
+            if (
+              statusChanges?.includes(NOTIFICATIONS_EVENT_NAMES.PROJECT_LISTED)
+            ) {
+              project.listed = true;
+              await project.save();
+            }
+
+            if (
+              statusChanges?.includes(
+                NOTIFICATIONS_EVENT_NAMES.PROJECT_UNLISTED,
+              )
+            ) {
+              project.listed = false;
+              await project.save();
+            }
+
+            if (
+              statusChanges?.includes(
+                NOTIFICATIONS_EVENT_NAMES.PROJECT_NOT_REVIEWED,
+              )
+            ) {
+              project.listed = null;
+              await project.save();
+            }
+
+            if (
+              statusChanges?.includes(
+                NOTIFICATIONS_EVENT_NAMES.PROJECT_CANCELLED,
+              )
+            ) {
+              await changeUserBoostingsAfterProjectCancelled({
+                projectId: project.id,
+              });
+            }
+          }
+          await Promise.all([
+            refreshUserProjectPowerView(),
+            refreshProjectFuturePowerView(),
+            refreshProjectPowerView(),
+          ]);
+          return request;
+        },
+      },
+      exportFilterToCsv: {
+        actionType: 'resource',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.EXPORT_FILTER_TO_CSV,
+          ),
+        handler: exportProjectsWithFiltersToCsv,
+        component: false,
+      },
+      listProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.LIST_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return listDelist(context, request, ReviewStatus.Listed);
+        },
+        component: false,
+      },
+      unlistProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.UNLIST_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return listDelist(context, request, ReviewStatus.NotListed);
+        },
+        component: false,
+      },
+      verifyProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.VERIFY_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return verifyProjects(context, request, true);
+        },
+        component: false,
+      },
+      rejectProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.REJECT_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return verifyProjects(context, request, false);
+        },
+        component: false,
+      },
+      // the difference is that it sends another segment event
+      revokeBadge: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.REVOKE_BADGE,
+          ),
+        handler: async (request, response, context) => {
+          return verifyProjects(context, request, false, true);
+        },
+        component: false,
+      },
+      activateProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.ACTIVATE_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return updateStatusOfProjects(context, request, ProjStatus.active);
+        },
+        component: false,
+      },
+      deactivateProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.DEACTIVATE_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return updateStatusOfProjects(context, request, ProjStatus.deactive);
+        },
+        component: false,
+      },
+      cancelProject: {
+        actionType: 'bulk',
+        isVisible: true,
+        isAccessible: ({ currentAdmin }) =>
+          canAccessProjectAction(
+            { currentAdmin },
+            ResourceActions.CANCEL_PROJECT,
+          ),
+        handler: async (request, response, context) => {
+          return updateStatusOfProjects(context, request, ProjStatus.cancelled);
+        },
+        component: false,
+      },
+    },
+  },
+};
